Login
Created login form. The form has a field for email and password. New profiles can't be created from the UI. Only by adding them manually to the data/users.txt file.

When the user logs in with a valid email and password they will be redirected to the information page. The page has placeholder text and a logout button in the upper corner.

The application has session handling so the browser remembers the user once logged in.